# docs/apifox/dist/gmaster-public.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "G-Master API - 公开接口文档",
    "version": "v0.13.2-gmaster.1",
    "description": "适合导入 Apifox 并对外发布的公开接口集合。\n\n包含：模型中转接口、用户侧令牌管理接口、系统状态与个人数据接口。\n生产地址：`https://gmapi.fun`\nOpenAI 兼容基址：`https://gmapi.fun/v1`\n后台接口基址：`https://gmapi.fun/api`\n仓库地址：`https://github.com/HereditaryDog/G-Master-API`"
  },
  "servers": [
    {
      "url": "https://gmapi.fun",
      "description": "G-Master API 生产环境"
    }
  ],
...
  "paths": {
    "/v1/models": {
      "get": {
        "summary": "获取模型列表",
        "deprecated": false,
        "description": "获取当前可用的模型列表。\n\n根据请求头自动识别返回格式：\n- 包含 `x-api-key` 和 `anthropic-version` 头时返回 Anthropic 格式\n- 包含 `x-goog-api-key` 头或 `key` 查询参数时返回 Gemini 格式\n- 其他情况返回 OpenAI 格式\n",
        "operationId": "listModels",
        "tags": [
          "获取模型列表"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Google API Key (用于 Gemini 格式)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Anthropic API Key (用于 Claude 格式)",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "anthropic-version",
            "in": "header",
            "description": "Anthropic API 版本",
            "required": false,
            "example": "",
            "schema": {
              "type": "string",
              "example": "2023-06-01"
            }
          },
          {
            "name": "x-goog-api-key",
            "in": "header",
            "description": "Google API Key (用于 Gemini 格式)",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                },
                "example": {
                  "object": "list",
                  "data": [
                    {
                      "id": "gpt-4o-mini",
                      "object": "model",
                      "owned_by": "g-master-api"
                    },
                    {
                      "id": "claude-3-5-sonnet-latest",
                      "object": "model",
                      "owned_by": "g-master-api"
                    }
                  ]
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1beta/models": {
      "get": {
        "summary": "Gemini 格式获取",
        "deprecated": false,
        "description": "以 Gemini API 格式返回可用模型列表",
        "operationId": "listModelsGemini",
        "tags": [
          "获取模型列表"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeminiModelsResponse"
                },
                "example": {
                  "models": [
                    {
                      "name": "models/gemini-2.5-flash",
                      "displayName": "Gemini 2.5 Flash"
                    },
                    {
                      "name": "models/gemini-2.5-pro",
                      "displayName": "Gemini 2.5 Pro"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/responses/compact": {
      "post": {
        "summary": "压缩对话 (OpenAI Responses API)",
        "deprecated": false,
        "description": "OpenAI Responses API，用于对长对话进行 compaction。",
        "operationId": "compactResponse",
        "tags": [
          "OpenAI格式(Responses)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponsesCompactionRequest"
              },
              "example": {
                "model": "gpt-4.1-mini",
                "input": "写一个简短的产品介绍。",
                "stream": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功压缩对话",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsesCompactionResponse"
                },
                "example": {
                  "id": "resp_demo_002",
                  "model": "gpt-4.1-mini",
                  "output_text": "这是一段简短的产品介绍。"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/images/generations": {
      "post": {
        "summary": "生成图像(qwen-image)",
        "deprecated": false,
        "description": " 百炼qwen-image系列图片生成",
        "operationId": "createImage",
        "tags": [
          "图片生成/Qwen千问"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "input": {
                    "type": "object",
                    "properties": {
                      "messages": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "role": {
                              "type": "string"
                            },
                            "content": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "text": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "required": [
                      "messages"
                    ]
                  },
                  "parameters": {
                    "type": "object",
                    "properties": {
                      "negative_prompt": {
                        "type": "string"
                      },
                      "prompt_extend": {
                        "type": "boolean"
                      },
                      "watermark": {
                        "type": "boolean"
                      },
                      "size": {
                        "type": "string"
                      }
                    }
                  }
                },
                "required": [
                  "model",
                  "input"
                ]
              },
              "example": {
                "model": "gpt-image-1",
                "prompt": "一张具有科技感的蓝紫色 AI 控制台首页横幅",
                "size": "1024x1024"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功生成图像",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                },
                "example": {
                  "created": 1775643000,
                  "data": [
                    {
                      "url": "https://example.com/generated-image.png",
                      "revised_prompt": "一张具有科技感的蓝紫色 AI 控制台首页横幅"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/images/edits": {
      "post": {
        "summary": "编辑图像(qwen-image-edit)",
        "deprecated": false,
        "description": " 百炼qwen-image系列图片生成",
        "operationId": "createImage",
        "tags": [
          "图片生成/Qwen千问"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "input": {
                    "type": "object",
                    "properties": {
                      "messages": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "role": {
                              "type": "string"
                            },
                            "content": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "image": {
                                    "type": "string"
                                  },
                                  "text": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "required": [
                      "messages"
                    ]
                  },
                  "parameters": {
                    "type": "object",
                    "properties": {
                      "n": {
                        "type": "integer"
                      },
                      "negative_prompt": {
                        "type": "string"
                      },
                      "prompt_extend": {
                        "type": "boolean"
                      },
                      "watermark": {
                        "type": "boolean"
                      },
                      "size": {
                        "type": "string"
                      }
                    }
                  }
                },
                "required": [
                  "model",
                  "input"
                ]
              },
              "example": {
                "model": "gpt-image-1",
                "prompt": "将背景改成蓝紫渐变，并增强科技感",
                "image": "@input.png"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功生成图像",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                },
                "example": {
                  "created": 1775643000,
                  "data": [
                    {
                      "url": "https://example.com/edited-image.png",
                      "revised_prompt": "将背景改成蓝紫渐变，并增强科技感"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/audio/transcriptions": {
      "post": {
        "summary": "音频转录",
        "deprecated": false,
        "description": "将音频转换为文本",
        "operationId": "createTranscription",
        "tags": [
          "OpenAI音频(Audio)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "音频文件",
                    "example": ""
                  },
                  "model": {
                    "type": "string",
                    "example": "whisper-1"
                  },
                  "language": {
                    "type": "string",
                    "description": "ISO-639-1 语言代码",
                    "example": ""
                  },
                  "prompt": {
                    "type": "string",
                    "example": ""
                  },
                  "response_format": {
                    "type": "string",
                    "enum": [
                      "json",
                      "text",
                      "srt",
                      "verbose_json",
                      "vtt"
                    ],
                    "default": "json",
                    "example": "json"
                  },
                  "temperature": {
                    "type": "number",
                    "example": 0
                  },
                  "timestamp_granularities": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "word",
                        "segment"
                      ]
                    },
                    "example": ""
                  }
                },
                "required": [
                  "file",
                  "model"
                ]
              },
              "example": {
                "file": "@speech.mp3",
                "model": "whisper-1",
                "language": "zh"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功转录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioTranscriptionResponse"
                },
                "example": {
                  "text": "欢迎使用 G-Master API。"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/audio/speech": {
      "post": {
        "summary": "文本转语音",
        "deprecated": false,
        "description": "将文本转换为音频",
        "operationId": "createSpeech",
        "tags": [
          "OpenAI音频(Audio)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechRequest"
              },
              "example": {
                "model": "gpt-4o-mini-tts",
                "input": "欢迎使用 G-Master API。",
                "voice": "alloy",
                "response_format": "mp3"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功生成音频",
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                },
                "example": "<binary audio stream>"
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "错误信息"
              },
              "type": {
                "type": "string",
                "description": "错误类型"
              },
              "param": {
                "type": "string",
                "description": "相关参数",
                "nullable": true
              },
              "code": {
                "type": "string",
                "description": "错误代码",
                "nullable": true
              }
            }
          }
        }
      },
...
      "ModelsResponse": {
        "type": "object",
        "properties": {
          "object": {
            "type": "string",
            "example": "list"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          }
        }
      },
...
      "AudioTranscriptionResponse": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session认证，通过登录接口获取"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Access Token认证，格式: Bearer {access_token}，通过 /api/user/token 接口生成"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "用户ID请求头，必须与当前登录用户ID匹配，使用Session或AccessToken认证时必须提供"
      },
      "Combination": {
        "group": [
          {
            "id": "SessionAuth"
          },
          {
            "id": "NewApiUser"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          },
          {
...
  }
}